reporting/order.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Reporting: Order</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
window.onload = function() {
  async_test(function(test) {
    var observer = new ReportingObserver(function(reports, observer) {
      test.step(function() {
        assert_equals(reports.length, 10);

        for(i = 0; i != 10; ++i)
          assert_equals(reports[i].body.message, "" + i);
      });

      test.done();
    });
    observer.observe();

    for (i = 0; i != 10; ++i) {
      let report = new CustomReport("" + i);
      report.send();
    }
  });
};
</script>
</body>
</html>
